// src/models/Enums/TypographyEnums.ts
export enum typographyConstants {
  SPECIAL_BOLD = "SPECIAL_BOLD",
  H1_BOLD = "H1_BOLD",
  H1_MEDIUM = "H1_MEDIUM",
  H1_REGULAR = "H1_REGULAR",
  H2_BOLD = "H2_BOLD",
  H2_MEDIUM = "H2_MEDIUM",
  H2_REGULAR = "H2_REGULAR",
  H3_BOLD = "H3_BOLD",
  H3_MEDIUM = "H3_MEDIUM",
  H3_REGULAR = "H3_REGULAR",
  H4_BOLD = "H4_BOLD",
  H4_MEDIUM = "H4_MEDIUM",
  H4_REGULAR = "H4_REGULAR",
  H5_BOLD = "H5_BOLD",
  H5_MEDIUM = "H5_MEDIUM",
  H5_REGULAR = "H5_REGULAR",
  H6_BOLD = "H6_BOLD",
  H6_MEDIUM = "H6_MEDIUM",
  H6_REGULAR = "H6_REGULAR",
  generic = 'generic'
}

export enum typographyWeight {
  regular = 'regular',
  medium = 'medium',
  bold = 'bold',
}

export enum typographyFamily {
  openSans = "Open Sans, sans-serif;",
}

export enum typographyColors {
  blue = "#002350",
  black = "#353535",
  white = "#FFFFFF",
}

export enum TextVariant {
  h1 = "h1",
  h2 = "h2",
  h3 = "h3",
  h4 = "h4",
  h5 = "h5",
  h6 = "h6",
}